requirements.txt
APScheduler==3.6.3
atomicwrites==1.4.0
attrs==21.2.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
flake8==3.9.2
flake8-docstrings==1.6.0
Flask==2.0.2
idna==3.3
importlib-metadata==4.8.2
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.7.0
pydocstyle==6.1.1
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.5
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
snowballstemmer==2.1.0
toml==0.10.2
tornado==6.1
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
